21/9/25 first 5 hours
today is a terrible day. It is dead dark, I am so tired but I have to keep coding. I had to restart my code because of my dubious implementation of CSS, resulting in a variety of unexpected errors, which made my code too far gone. I have implemented the start of a grid and have implemented photos and a heading.