3 hours later...
The clock strikes midnight. I have finished the first, and arguably the hardest page because of the amount of CSS that was needed. I added the 2 albums, and critic quotes beneath them. Also added a cool red button that takes the user to the next page and the more important page which is about local bands, their struggles backed by statistics, why you should support them and how to support them. By the way no caffeine has been abused here I promise, this is being done of pure sheer willpower. See you in the next update, whenever that is.