1 hour later
A lot of progress has been made. I have successfully coded the entire information section for the first band, Shiva, by incorporating CSS and HTML. I have added an interactable image which leads you from the main page to the information page, which then has some simple text about the bands influences and so on, and then a go back button which brings you back to the main page reliably. Next, I will be doing the 2 other pages and that will likely mark the completion of this website. Peace out until next time!